+++ allauth/socialaccount/providers/base/provider.py
@@ -170,7 +170,7 @@ class ProviderAccount(object):
                     dflt = super(GoogleAccount, self).__str__()
                     return self.account.extra_data.get('name', dflt)
 
-        So we have this method `to_str` that can be overriden in a conventional
+        So we have this method `to_str` that can be overridden in a conventional
         fashion, without having to worry about it.
         """
         return self.get_brand()["name"]